Functions

Functions

List of all functions in PolyChaos.

LinearAlgebra.issymmetric

PolyChaos.calculateAffinePCE

PolyChaos.clenshaw_curtis

PolyChaos.coeffs

PolyChaos.computeSP

PolyChaos.computeSP2

PolyChaos.convert2affinePCE

PolyChaos.evaluate

PolyChaos.evaluatePCE

PolyChaos.fejer

PolyChaos.fejer2

PolyChaos.gauss

PolyChaos.integrate

PolyChaos.lanczos

PolyChaos.lobatto

PolyChaos.lobatto_jacobi

PolyChaos.mcdiscretization

PolyChaos.nw

PolyChaos.quadgp

PolyChaos.r_scale

PolyChaos.radau

PolyChaos.radau_jacobi

PolyChaos.radau_laguerre

PolyChaos.rm_compute

PolyChaos.rm_hermite

PolyChaos.rm_hermite_prob

PolyChaos.rm_jacobi

PolyChaos.rm_jacobi01

PolyChaos.rm_laguerre

PolyChaos.rm_legendre

PolyChaos.rm_legendre01

PolyChaos.rm_logistic

PolyChaos.rm_meixner_pollaczek

PolyChaos.sampleMeasure

PolyChaos.samplePCE

PolyChaos.stieltjes

Statistics.mean

Statistics.std

Statistics.var

Recurrence Coefficients for Monic Orthogonal Polynomials

The functions below provide analytic expressions for the recurrence coefficients of common orthogonal polynomials. All of these provide monic orthogonal polynomials relative to the weights.

Note

The number N of recurrence coefficients has to be positive for all functions below.

PolyChaos.r_scale — Method.

r_scale(c::Float64,β::Vector{Float64},α::Vector{Float64})

Given the recursion coefficients (α,β) for a system of orthogonal polynomials that are orthogonal with respect to some positive weight $m(t)$, this function returns the recursion coefficients (α_,β_) for the scaled measure $c m(t)$ for some positive $c$.

source

PolyChaos.rm_compute — Method.

rm_compute(weight::Function,lb::Float64,ub::Float64;Npoly::Int64=4,Nquad::Int64=10,quadrature::Function=clenshaw_curtis)

Given a positive weight function with domain (lb,ub), i.e. a function $w: [lb, ub ] \rightarrow \mathbb{R}_{\geq 0}$, this function creates Npoly recursion coefficients (α,β).

The keyword quadrature specifies what quadrature rule is being used.

source

PolyChaos.rm_logistic — Method.

rm_logistic(N::Int)

Creates N recurrence coefficients for monic polynomials that are orthogonal on $(-\infty,\infty)$ relative to $w(t) = \frac{\mathrm{e}^{-t}}{(1 - \mathrm{e}^{-t})^2}$

source

PolyChaos.rm_hermite — Method.

rm_hermite(N::Int,mu::Float64)
rm_hermite(N::Int)

Creates N recurrence coefficients for monic generalized Hermite polynomials that are orthogonal on $(-\infty,\infty)$ relative to $w(t) = |t|^{2 \mu} \mathrm{e}^{-t^2}$

The call rm_hermite(N) is the same as rm_hermite(N,0).

source

PolyChaos.rm_hermite_prob — Method.

rm_hermite_prob(N::Int)

Creates N recurrence coefficients for monic probabilists' Hermite polynomials that are orthogonal on $(-\infty,\infty)$ relative to $w(t) = \mathrm{e}^{-0.5t^2}$

source

PolyChaos.rm_laguerre — Method.

rm_laguerre(N::Int,a::Float64)
rm_laguerre(N::Int)

Creates N recurrence coefficients for monic generalized Laguerre polynomials that are orthogonal on $(0,\infty)$ relative to $w(t) = t^a \mathrm{e}^{-t}$.

The call rm_laguerre(N) is the same as rm_laguerre(N,0).

source

PolyChaos.rm_legendre — Method.

rm_legendre(N::Int)

Creates N recurrence coefficients for monic Legendre polynomials that are orthogonal on $(-1,1)$ relative to $w(t) = 1$.

source

PolyChaos.rm_legendre01 — Method.

rm_legendre01(N::Int)

Creates N recurrence coefficients for monic Legendre polynomials that are orthogonal on $(0,1)$ relative to $w(t) = 1$.

source

PolyChaos.rm_jacobi — Method.

rm_jacobi(N::Int,a::Float64,b::Float64)
rm_jacobi(N::Int,a::Float64)
rm_jacobi(N::Int)

Creates N recurrence coefficients for monic Jacobi polynomials that are orthogonal on $(-1,1)$ relative to $w(t) = (1-t)^a (1+t)^b$.

The call rm_jacobi(N,a) is the same as rm_jacobi(N,a,a) and rm_jacobi(N) the same as rm_jacobi(N,0,0).

source

PolyChaos.rm_jacobi01 — Method.

rm_jacobi01(N::Int,a::Float64,b::Float64)
rm_jacobi01(N::Int,a::Float64)
rm_jacobi01(N::Int)

Creates N recurrence coefficients for monic Jacobi polynomials that are orthogonal on $(0,1)$ relative to $w(t) = (1-t)^a t^b$.

The call rm_jacobi01(N,a) is the same as rm_jacobi01(N,a,a) and rm_jacobi01(N) the same as rm_jacobi01(N,0,0).

source

PolyChaos.rm_meixner_pollaczek — Method.

rm_meixner_pollaczek(N::Int,lambda::Float64,phi::Float64)
rm_meixner_pollaczek(N::Int,lambda::Float64)

Creates N recurrence coefficients for monic Meixner-Pollaczek polynomials with parameters λ and ϕ. These are orthogonal on $[-\infty,\infty]$ relative to the weight function $w(t)=(2 \pi)^{-1} \exp{(2 \phi-\pi)t} |\Gamma(\lambda+ i t)|^2$.

The call rm_meixner_pollaczek(n,lambda) is the same as rm_meixner_pollaczek(n,lambda,pi/2).

source

PolyChaos.stieltjes — Function.

stieltjes(N::Int64,nodes_::Vector{Float64},weights_::Vector{Float64};removezeroweights::Bool=true)

Stieltjes procedure–-Given the nodes and weights the function generates the firstN recurrence coefficients of the corresponding discrete orthogonal polynomials.

Set the Boolean removezeroweights to true if zero weights should be removed.

source

PolyChaos.lanczos — Function.

lanczos(N::Int64,nodes_::Vector{Float64},weights_::Vector{Float64};removezeroweights::Bool=true)

Lanczos procedure–-given the nodes and weights the function generates the first N recurrence coefficients of the corresponding discrete orthogonal polynomials.

Set the Boolean removezeroweights to true if zero weights should be removed.

The script is adapted from the routine RKPW in W.B. Gragg and W.J. Harrod, The numerically stable reconstruction of Jacobi matrices from spectral data, Numer. Math. 44 (1984), 317-335.

source

PolyChaos.mcdiscretization — Function.

mcdiscretization(N::Int64,quads::Vector{},discretemeasure::Matrix{Float64}=zeros(0,2);discretization::Function=stieltjes,Nmax::Integer=300,ε::Float64=1e-8,gaussquad::Bool=false)

This routine returns $N$ recurrence coefficients of the polynomials that are orthogonal relative to a weight function $w$ that is decomposed as a sum of $m$ weights $w_i$ with domains $[a_i,b_i]$ for $i=1,\dots,m$,

\[w(t) = \sum_{i}^{m} w_i(t) \quad \text{with } \operatorname{dom}(w_i) = [a_i, b_i].\]

For each weight $w_i$ and its domain $[a_i, b_i]$ the function mcdiscretization() expects a quadrature rule of the form nodes::Vector{Float64}, weights::Vector{Float64} = myquadi(N::Int64) all of which are stacked in the parameter quad quad = [ myquad1, ..., myquadm ] If the weight function has a discrete part (specified by discretemeasure) it is added on to the discretized continuous weight function.

The function mcdiscretization() performs a sequence of discretizations of the given weight $w(t)$, each discretization being followed by an application of the Stieltjes or Lanczos procedure (keyword discretization in [stieltjes, lanczos]) to produce approximations to the desired recurrence coefficients. The function applies to each subinterval $i$ an N-point quadrature rule (the $i$th entry of quad) to discretize the weight function $w_i$ on that subinterval. If the procedure converges to within a prescribed accuracy ε before N reaches its maximum allowed value Nmax. If the function does not converge, the function prompts an error message.

The keyword gaussquad should be set to true if Gauss quadrature rules are available for all $m$ weights $w_i(t)$ with $i = 1, \dots, m$.

For further information, please see W. Gautschi "Orthogonal Polynomials: Approximation and Computation", Section 2.2.4.

source

Scalar Products

PolyChaos.computeSP2 — Function.

computeSP2(n::Int64,β::Vector{Float64})
computeSP2(n::Int64,op::OrthoPoly) = computeSP2(n,op.β)
computeSP2(op::OrthoPoly) = computeSP2(op.deg,op.β)
computeSP2(opq::OrthoPolyQ) = computeSP2(opq.op)

Computes the n regular scalar products aka 2-norms of the orthogonal polynomials, namely

\[\|ϕ_i\|^2 = \langle \phi_i,\phi_i\rangle \quad \forall i \in \{ 0,\dots,n \}.\]

Notice that only the values of β of the recurrence coefficients (α,β) are required. The computation is based on equation (1.3.7) from Gautschi, W. "Orthogonal Polynomials: Computation and Approximation". Whenever there exists an analytic expressions for β, this function should be used.

The function is multiply dispatched to facilitate its use with the composite types OrthoPoly and OrthoPolyQ.

source

PolyChaos.computeSP — Function.

Univariate

computeSP(a::Vector{Int64},α::Vector{Float64},β::Vector{Float64},nodes::Vector{Float64},weights::Vector{Float64};issymmetric::Bool=false)
computeSP(a::Vector{Int64},op::OrthoPoly,q::Quad;issymmetric=issymmetric(op))
computeSP(a::Vector{Int64},opq::OrthoPolyQ)

Multivariate

computeSP( a::Vector{Int64},
           α::Vector{Vector{Float64}},β::Vector{Vector{Float64}},
           nodes::Vector{Vector{Float64}},weights::Vector{Vector{Float64}},
           ind::Matrix{Int64};
           issymmetric::BitArray=falses(length(α)))
computeSP(a::Vector{Int64},op::Vector{OrthoPoly},q::Vector{Quad},ind::Matrix{Int64})
computeSP(a::Vector{Int64},mOP::MultiOrthoPoly)

Computes the scalar product

\[\langle \phi_{a_1},\phi_{a_2},\cdots,\phi_{a_n} \rangle,\]

where n = length(a). This requires to provide the recurrence coefficients (α,β) and the quadrature rule (nodes,weights), as well as the multi-index ind. If provided via the keyword issymmetric, symmetry of the weight function is exploited. All computations of the multivariate scalar products resort back to efficient computations of the univariate scalar products. Mathematically, this follows from Fubini's theorem.

The function is multiply dispatched to facilitate its use with OrthoPolyQ or a suitable combination of OrthoPoly and its quadrature rule Quad.

Note

Zero entries of $a$ are removed automatically to simplify computations, which follows from

\[\langle \phi_i, \phi_j, \phi_0,\cdots,\phi_0 \rangle = \langle \phi_i, \phi_j \rangle,\]

because \phi_0 = 1.

It is checked whether enough quadrature points are supplied to solve the integral exactly.

source

Evaluating Orthogonal Polynomials

PolyChaos.evaluate — Function.

Univariate

evaluate(n::Int64,x::Array{Float64},a::Vector{Float64},b::Vector{Float64})
evaluate(n::Int64,x::Float64,a::Vector{Float64},b::Vector{Float64})
evaluate(n::Int64,x::Vector{Float64},op::OrthoPoly)
evaluate(n::Int64,x::Float64,op::OrthoPoly)

Evaluate the n-th univariate basis polynomial at point(s) x The function is multiply dispatched to facilitate its use with the composite type OrthoPoly

Note

n is the degree of the univariate basis polynomial

length(x) = N, where N is the number of points

(a,b) are the recursion coefficients

Multivariate

evaluate(n::Vector{Int64},x::Matrix{Float64},a::Vector{Vector{Float64}},b::Vector{Vector{Float64}})
evaluate(n::Vector{Int64},x::Vector{Float64},a::Vector{Vector{Float64}},b::Vector{Vector{Float64}})
evaluate(n::Vector{Int64},x::Matrix{Float64},op::MultiOrthoPoly)
evaluate(n::Vector{Int64},x::Vector{Float64},op::MultiOrthoPoly)

Evaluate the n-th p-variate basis polynomial at point(s) x The function is multiply dispatched to facilitate its use with the composite type MultiOrthoPoly

Note

`n is a multi-index

length(n) == p, i.e. a p-variate basis polynomial

size(x) = (N,p), where N is the number of points

size(a)==size(b)=p.

source

Quadrature Rules

PolyChaos.fejer — Function.

fejer(N::Int64)

Fejer's first quadrature rule.

source

PolyChaos.fejer2 — Function.

fejer2(n::Int64)

Fejer's second quadrature rule according to Waldvogel, J. Bit Numer Math (2006) 46: 195.

source

PolyChaos.clenshaw_curtis — Function.

clenshaw_curtis(n::Int64)

Clenshaw-Curtis quadrature according to Waldvogel, J. Bit Numer Math (2006) 46: 195.

source

PolyChaos.quadgp — Function.

quadgp(weight::Function,a::Float64,b::Float64,N::Int64=10;quadrature::Function=clenshaw_curti

general purpose quadrature based on Gautschi, "Orthogonal Polynomials: Computation and Approximation", Section 2.2.2, pp. 93-95

source

PolyChaos.gauss — Function.

gauss(N::Int64,α::Vector{Float64},β::Vector{Float64})
gauss(α::Vector{Float64},β::Vector{Float64})
gauss(N::Int64,op::OrthoPoly)
gauss(op::OrthoPoly)

Gauss quadrature rule, also known as Golub-Welsch algorithm

gauss() generates the N Gauss quadrature nodes and weights for a given weight function. The weight function is represented by the N recurrence coefficients for the monic polynomials orthogonal with respect to the weight function.

Note

If no N is provided, then N = length(α).

source

PolyChaos.radau — Function.

radau(N::Int64,α::Vector{Float64},β::Vector{Float64},end0::Float64)
radau(α::Vector{Float64},β::Vector{Float64},end0::Float64)
radau(N::Int64,op::OrthoPoly,end0::Float64)
radau(op::OrthoPoly,end0::Float64)

Gauss-Radau quadrature rule. Given a weight function encoded by the recurrence coefficients (α,β)for the associated orthogonal polynomials, the function generates the nodes and weights (n+1)-point Gauss-Radau quadrature rule for the weight function having a prescribed node end0 (typically at one of the end points of the support interval of w, or outside thereof).

Note

If no N is specified, then N = length(α)-1.

Note

Reference: OPQ: A MATLAB SUITE OF PROGRAMS FOR GENERATING ORTHOGONAL POLYNOMIALS AND RELATED QUADRATURE RULES by Walter Gautschi

source

PolyChaos.radau_jacobi — Function.

radau_jacobi(N::Int64,a::Float64,b::Float64;endpoint::String="left")
endpoint in ["left", "right"]
radau_jacobi(N::Int64,a::Float64;endpoint::String="left") = radau_jacobi(N,a,a;endpoint=endpoint)
radau_jacobi(N::Int64;endpoint::String="left") = radau_jacobi(N,0.;endpoint=endpoint)

Gauss-Radau quadrature rule for Jacobi weight function, which generates the (n+1)-point Gauss-Radau rule for the Jacobi weight function on [-1,1] with parameters a and b.

Note

REFERENCE: W. Gautschi, ``Gauss-Radau formulae for Jacobi and Laguerre weight functions'', Math. Comput. Simulation 54 (2000), 403-412.

source

PolyChaos.radau_laguerre — Function.

radau_laguerre(N::Int64,a::Float64)
radau_laguerre(N::Int64) = radau_laguerre(N,0.)

Gauss-Radau quadrature rule for Laguerre weight function, which generates the (n+1)-point Gauss-Radau rule for the Laguerre weight function on $[0,\infty]$ with parameter a.

Note

REFERENCE: W. Gautschi, ``Gauss-Radau formulae for Jacobi and Laguerre weight functions'', Math. Comput. Simulation 54 (2000), 403-412.

source

PolyChaos.lobatto — Function.

lobatto(N::Int64,α::Vector{Float64},β::Vector{Float64},endl::Float64,endr::Float64)
lobatto(α::Vector{Float64},β::Vector{Float64},endl::Float64,endr::Float64)
lobatto(N::Int64,op::OrthoPoly,endl::Float64,endr::Float64)
lobatto(op::OrthoPoly,endl::Float64,endr::Float64)

Gauss-Lobatto quadrature rule. Given a weight function encoded by the recurrence coefficients for the associated orthogonal polynomials, the function generates the nodes and weights of the (N+2)-point Gauss-Lobatto quadrature rule for the weight function, having two prescribed nodes endl, endr (typically the left and right end points of the support interval, or points to the left resp. to the right therof).

Note

If no N is specified, then N = length(α)-2.

Note

Reference: OPQ: A MATLAB SUITE OF PROGRAMS FOR GENERATING ORTHOGONAL POLYNOMIALS AND RELATED QUADRATURE RULES by Walter Gautschi

source

PolyChaos.lobatto_jacobi — Function.

lobatto_jacobi(N::Int64,a::Float64,b::Float64)
lobatto_jacobi(N::Int64,a::Float64) = lobatto_jacobi(N,a,a)
lobatto_jacobi(N::Int64) = lobatto_jacobi(N,0.)

Gauss-Lobatto quadrature rule for Jacobi weight function, which generates the (n+2)-point Gauss- Lobatto rule for the Jacobi weight function on $[-1,1]$ with parameters a and b.

Note

REFERENCE: W. Gautschi,``High-order Gauss-Lobatto formulae'', Numer. Algorithms 25 (2000), 213-222.

source

Polynomial Chaos

Statistics.mean — Function.

Univariate

mean(x::Vector{Float64},op::OrthoPoly)
mean(x::Vector{Float64},opq::OrthoPolyQ)

Multivariate

mean(x::Vector{Float64},mop::MultiOrthoPoly)

compute mean of random variable with PCE x

source

Statistics.var — Function.

Univariate

var(x::Vector{Float64},op::OrthoPoly)
var(x::Vector{Float64},opq::OrthoPolyQ)

Multivariate

var(x::Vector{Float64},mop::MultiOrthoPoly)

compute variance of random variable with PCE x

source

Statistics.std — Function.

Univariate

std(x::Vector{Float64},op::OrthoPoly)
std(x::Vector{Float64},opq::OrthoPolyQ)

Multivariate

std(x::Vector{Float64},mop::MultiOrthoPoly)

compute standard deviation of random variable with PCE x

source

PolyChaos.sampleMeasure — Function.

Univariate

sampleMeasure(n::Int64,name::String,w::Function,dom::Tuple{Float64,Float64},symm::Bool,d::Dict;method::String="adaptiverejection")
sampleMeasure(n::Int64,m::Measure;method::String="adaptiverejection")
sampleMeasure(n::Int64,op::OrthoPoly;method::String="adaptiverejection")
sampleMeasure(n::Int64,opq::OrthoPolyQ;method::String="adaptiverejection")

Draw n samples from the measure m described by its

name

weight function w,

domain dom,

symmetry property symm,

and, if applicable, parameters stored in the dictionary d.

By default an adaptive rejection sampling method is used (from AdaptiveRejectionSampling.jl), unless it is a common random variable for which Distributions.jl is used.

The function is multiply dispatched to accept OrthoPoly or OrthoPolyQ.

Multivariate

sampleMeasure(n::Int64,m::MultiMeasure;method::Vector{String}=["adaptiverejection" for i=1:length(m.name)])
sampleMeasure(n::Int64,mop::MultiOrthoPoly;method::Vector{String}=["adaptiverejection" for i=1:length(mop.meas.name)])

Multivariate extension which provides array of samples with n rows and as many columns as the multimeasure has univariate measures.

source

PolyChaos.evaluatePCE — Function.

evaluatePCE(x::Vector{Float64},ξ::Vector{Float64},α::Vector{Float64},β::Vector{Float64})

Evaluation of polynomial chaos expansion

\[\mathsf{x} = \sum_{i=0}^{L} x_i \phi_i{\xi_j},\]

where L+1 = length(x) and $x_j$ is the $j$th sample where $j=1,\dots,m$ with m = length(ξ).

source

PolyChaos.samplePCE — Function.

Univariate

samplePCE(n::Int64,x::Vector{Float64},op::OrthoPoly;method::String="adaptiverejection")
samplePCE(n::Int64,x::Vector{Float64},opq::OrthoPolyQ;method::String="adaptiverejection")

Combines sampleMeasure and evaluatePCE, i.e. it first draws n samples from the measure, then evaluates the PCE for those samples.

Multivariate

samplePCE(n::Int64,x::Vector{Float64},mop::MultiOrthoPoly;method::Vector{String}=["adaptiverejection" for i=1:length(mop.meas.name)])

source

PolyChaos.calculateAffinePCE — Function.

calculateAffinePCE(α::Vector{Float64})::Vector{Float64}

Computes the affine PCE coefficients $x_0$ and $x_1$ from recurrence coefficients $lpha$.

source

PolyChaos.convert2affinePCE — Function.

convert2affinePCE(a::Vector{Float64},α0::Float64)
convert2affinePCE(name::String,p1::Float64,p2::Float64,d::Dict=Dict();kind::Symbol=:lbub)

Computes the affine PCE coefficients $x_0$ and $x_1$ from

\[X = a_1 + a_2 \Xi = x_0 + x_1 \phi_1(\Xi),\]

where $\phi_1(t) = t-\alpha_0$ is the first-order monic basis polynomial.

For classical polynomials the name can be given directly. The keyword kind in [:lbub, :μσ] specifies whether p1 and p2 have the meaning of lower/upper bounds or mean/standard deviation.

source

Auxiliary Functions

PolyChaos.nw — Function.

nw(q::Quad)
nw(opq::OrthoPolyQ)
nw(opq::Vector{OrthoPolyQ})
nw(mOP::MultiOrthoPoly)

returns nodes and weights in matrix form

source

PolyChaos.coeffs — Function.

coeffs(op::OrthoPoly)
coeffs(opq::OrthoPolyQ)
coeffs(op::Vector{OrthoPoly})
coeffs(opq::Vector{OrthoPolyQ})
coeffs(mop::MultiOrthoPoly)

returns recurrence coefficients of in matrix form

source

PolyChaos.integrate — Function.

integrate(f::Function,nodes::Vector{Float64},weights::Vector{Float64})
integrate(f::Function,q::Quad)
integrate(f::Function,opq::OrthogonalPolyQ)

integrate function f using quadrature rule specified via nodes, weights

Note

function $f$ is assumed to return a scalar

interval of integration is "hidden" in $nodes$

source

LinearAlgebra.issymmetric — Function.

issymmetric(m::Measure)::Bool
issymmetric(op::OrthoPoly)::Bool
issymmetric(q::Quad)::Bool
issymmetric(opq::OrthoPolyQ)::Bool

is measure symmetric (around any point in the domain)?

source